refactor overlay into separate module
