+++ src/frontend/OverlayManager/types.ts
@@ -1,0 +1,6 @@
+import { Runner } from 'common/types'
+
+export interface BrowserGameProps {
+  appName: string
+  runner: Runner
+}
